requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
langchain-openai
langchain-qdrant
qdrant-client
pypdf
tiktoken